pidx: force checking pdsc files
This patch adds support for checking pdsc URL, Vendor, Name or Version
information against the actual pdsc file.

Setting `PidxXML.force` to true will make AddPdsc download the pdsc file
and validate if what's in the pdsc tag matches the info in the file itself.

When it matches, then things are OK, when it doesn't match, raise an error
and add the pdsc from the file, instead of the tag.

When pdsc files aren not reachable, raise a error saying the reason why, then
adds the pdsc tag as if `.force` was set to False.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>